requirements.txt
discord.py==2.3.1
numpy==1.25.1
PyNaCl==1.5.0
ffmpeg==1.4
yt_dlp==2023.7.6
python-dotenv==1.0.0
